Correct PIMPL example and test
Thanks to Bengt Gustaffson for pointing this out.
