+++ BUILD.bazel
@@ -134,7 +134,10 @@ cc_binary(
 
 cc_library(
     name = "indirect_pimpl",
-    srcs = ["indirect_pimpl.cc"],
+    srcs = [
+        "indirect_pimpl.cc",
+        "indirect_pimpl_use.cc",
+    ],
     hdrs = ["indirect_pimpl.h"],
     deps = ["indirect"],
 )
@@ -146,7 +149,10 @@ build_test(
 
 cc_library(
     name = "polymorphic_pimpl",
-    srcs = ["polymorphic_pimpl.cc"],
+    srcs = [
+        "polymorphic_pimpl.cc",
+        "polymorphic_pimpl_use.cc",
+    ],
     hdrs = ["polymorphic_pimpl.h"],
     deps = ["polymorphic"],
 )
