+++ polymorphic_pimpl.cc
@@ -18,14 +18,11 @@ ClassWithPolymorphicPimpl::ClassWithPolymorphicPimpl(
 ClassWithPolymorphicPimpl& ClassWithPolymorphicPimpl::operator=(
     const ClassWithPolymorphicPimpl&) = default;
 
+ClassWithPolymorphicPimpl::ClassWithPolymorphicPimpl(
+    ClassWithPolymorphicPimpl&&) noexcept = default;
+
+ClassWithPolymorphicPimpl& ClassWithPolymorphicPimpl::operator=(
+    ClassWithPolymorphicPimpl&&) noexcept = default;
+
 void ClassWithPolymorphicPimpl::do_something() { impl_->do_something(); }
 }  //  namespace xyz::testing
-
-void check_methods() {
-  xyz::testing::ClassWithPolymorphicPimpl c; // Default construction.
-  auto cc = c; // Copy construction.
-  auto mc = std::move(c); // Move construction.
-  c = cc; // Copy assignment.
-  c = xyz::testing::ClassWithPolymorphicPimpl(); // Move assignment.
-  c.do_something();
-}
